# pulsesend-api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PulseSend Email Campaign Platform API",
    "description": "Official API Specification for PulseSend. Provides administrative control, contact handling, template management, analytics harvesting, and multi-tenant isolation hooks.",
    "version": "1.0.0",
    "contact": {
      "name": "PulseSend Engineering",
      "email": "dev@pulsesend.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Sandbox"
    },
    {
      "url": "https://api.pulsesend.com",
      "description": "Production API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/send-test": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Fire Solo Live Diagnostic Despatch",
        "description": "Direct, non-pipeline AWS SES execution allowing super-administrators to fire physical debug emails outside normal reporting workflows.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "to",
                  "subject"
                ],
                "properties": {
                  "to": {
                    "type": "string",
                    "format": "email"
                  },
                  "subject": {
                    "type": "string"
                  },
                  "html": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fired via AWS SES SMTP gateway"
          }
        }
      }
    },
...
  }
}